Remove overview section
I have decided to remove the overview section, because I don't want
this project to be a full guide into dart date formatting at this time.
It should just be a playground with all the useful information to format
dates. Short explainations of several things will be put into the
subsections.